requirements.txt
openai
langchain
langchain-openai
langchain-google-genai
python-dotenv